organize code for testing
