+++ src/static-methods.js
